0823 오후 TIL

+++ spring-practice/blog_proj/src/main/java/com/kosta/service/BlogServiceImpl.java
@@ -6,6 +6,7 @@ import java.util.Optional;
 import org.springframework.stereotype.Service;
 
 import com.kosta.entity.Article;
+import com.kosta.entity.User;
 import com.kosta.repository.BlogRepository;
 
 import lombok.RequiredArgsConstructor;
@@ -16,7 +17,8 @@ public class BlogServiceImpl implements BlogService {
 	private final BlogRepository blogRepository;
 
 	@Override
-	public Article save(Article article) {
+	public Article save(Article article, User user) {
+		article.setCreator(user);
 		return blogRepository.save(article);
 	}
 
@@ -33,14 +35,25 @@ public class BlogServiceImpl implements BlogService {
 	}
 
 	@Override
-	public void deleteById(Long id) throws Exception {
+	public void deleteById(Long id, User user) throws Exception {
 		Article article = findById(id);
-		blogRepository.deleteById(article.getId());
+		User creator = article.getCreator(); // creator : 작성자 / user : 로그인한 사용자
+		if (user.getId().equals(creator.getId())) {
+			// 삭제 가능
+			blogRepository.deleteById(article.getId());
+		} else {
+			throw new Exception("본인이 작성한 글만 삭제할 수 있습니다.");
+		}
 	}
 
 	@Override
-	public Article update(Article article) throws Exception {
+	public Article update(Article article, User user) throws Exception {
 		Article originArticle = findById(article.getId());
+		User creator = originArticle.getCreator();
+		if (!user.getId().equals(creator.getId())) {
+			// 수정 권한 없음
+			throw new Exception("본인이 작성한 글만 수정할 수 있습니다.");
+		}
 		originArticle.setTitle(article.getTitle());
 		originArticle.setContent(article.getContent());
 
@@ -58,4 +71,5 @@ public class BlogServiceImpl implements BlogService {
 		// 오름차순 결과 반환
 		return blogRepository.findByTitleContainsOrContentContainsOrderByTitleAsc(keyword, keyword);
 	}
+
 }
